orgenize code and add comments
